Improve exception reporting
Previously it could have happend, that no exception was available because
there was none in the result and the catched exception was not reported.

To fix this, we initialize the exception to report with the catched one.
When the result also has an exception it will override the default one
anyway.