Optimize assertion handling
When a AssertionError was thrown, it was not tracked by the assertThatSoon
call. So we missed the original error in the stack trace.
After this change, we don't need to catch AssertionErrors in the keepTrying
call anymore, as assertions only happen in assertThatSoon calls.

Additionally we add a assertThatLater method that is a shortcut for
assertThatSonn(getLongTimeoutInSeconds(), ...)