Removed callWithTimeout from SeleniumUtils.keepTrying
Keep trying is used really a lot. It contains a call to
SeleniumUtils.callWithTimeout, that uses an own thread to detect endless
loops. I removed this for multiple reasons:
* Endless loops must not happen. They must not happen with regular tests
and must not happen with integration tests. Why should we try to detect
those?
* The timeout of the keepTrying still works, it's not affected by the
change.
* Stack-Traces are much harder to interpret because parts of the
transaction run in a seperate transaction.
* It causes a lag.
* It's simply not needed :)