Shorter, more precise exception message in assertThatSoon/Later
The assert was flooding the log with exceptions. The reason for the
initial change, was, that the assert has not been logged. But the reason
for this was, that the assert did just return "false" but did not throw
and exception. The code did not solve this at all, but flooded the
log with exceptions.

The current change hopefully solves this (and did in all tested
scenarios)