Separeted the AbstractSeleniumPage
There is an AbstractSeleniumStandalonePage, now, that contains
the open and url methods. The problem is, that a Page must not be
able to open itself. E.g. if you click on a search result and open the
details, this details page will not be loaded by URL, but rather
just represent the "current" page of the envrionment.

This was not possible. The "standalone" page, now contains the
old functionality with a mandatory URL and an open method.