Reenable KeepTrying within KeepTrying
This influences the assertThatSoon/Later calls. Until now any keepTrying
call within another keepTrying was ignored (tried only once). This was
done, in order to tackle some timeout problems with calls that used a
longer timeout within a shorter timeout.

But this caused other problems.

E.g. if you search for results and pick a specific one. As the search index
hangs behind, it is necessary to call refresh until you find the specific
result. But refreshing takes time, too. In order to fix this, you make
an outer assertThatLater for the result and an inner assertThatSoon
for any results to appear. But the inner assert was ignored, causing
the code to constantly refresh without waiting for the results to appear.
There was no way to fix this, unless removing this block.

Running local tests, this fixes some other timing problems, too.