Add highlight method to component
Sometimes, when a selector matches multiple elements on screen, it is
hard to find if the right element was selected.

The new highlight method makes this easier. You can debug up to the point,
where the problem occours, and call the highlight function during the
debug session. Now the element in question gets a background color and
border. So it should be easy to see if it is the correct one.