Fix selecting/deselecting Clarity checkboxes without label
We now click on the whole Clarity checkbox in
order to select/deselect it instead of the label,
because the label may be empty.

Clicking on the label would be a good test, indeed,
but we can assume, that Clarity is doing this right
and correctly assigns the label to the box.