Only consider `ActiveSeleniumComponent`s with dimensions > 0 visible
Reason for this change is that around v7.10.0, Clarity changed the size
of empty checkbox label tags: while previously they had a padding that
caused them to be size a * b, even without any content, the padding was
now removed and the resulting size is 0 * b. This causes an
`ElementNotInteractableException: element not interactable: element has zero size`
when Selenium attempts to click on the label.

Therefore, the `ActiveSeleniumComponent` interface now overrides the
`isVisible` method with a custom implementation that checks if both x
and y dimensions are greater than 0.