Create `ActiveSeleniumComponent#isInteractable`
Reason for this change is that around v7.10, Clarity changed the size
of empty checkbox label tags: while previously they had a padding that
caused them to be size a * b, even without any content, the padding was
now removed and the resulting size is 0 * b. This causes an
`ElementNotInteractableException: element not interactable: element has zero size`
when Selenium attempts to click on the label.

So a method was added to be able to check if the element can be
interacted with before the interaction is attempted.

(I would have made this part of the #isVisible or #isClickable method,
but that causes unsolvable problems with selecting `<option>`s in the
`SelectComponent`.)